Update event system
EventInfo has been reverted to a regular c# class so instances can be created in code.

Added EventInfoMono. This contains EventInfo and allows ease of creating and assigning EventInfo in the Inspector.

Added EventInfoSO. This contains EventInfo and allows creation of persistent, cross-scene EventInfo that can be easily assigned in the inspector..