Add game board, tiles and related content
Added tiles that can be of a certain TileType.

Added initial textures, materials and TileTypes.

Added BoardManager that handles creating the board out of random tiles with a given percent of each.

Added TileBorders that creates borders around tiles in a given area. The borders are a VFX.

Added ReadOnly Attribute and PropertyDrawer that makes fields visible in the Inspector readonly.

Added an Event for when a tile is selected (onDownUp interaction).

Updated the SampleSceneMonoInstaller to account the required dependencies.