Update
Changed the BoardManager.boardTileWidth and .boardTileHeight to .widthInTiles and .heightInTiles respectively, also made them non-static so they can be drawn in the Inspector and can be access cross-scene with Zenject.

Made BoardManager.tiles non-static and a property so it can be accessed cross-scene with Zenject and can't be changed outside of the class.

Added a method to BoardManager to try and get a tile, returns null if the given position is out of range. This just removes the need for classes to check themselves, avoiding repeated code.