Add EventConstraints support
Added support for event constraints. If an event is invoked it will only call it's subscribers if their constraints match.

EventConstraints currently include GameState and the EventInfo's gameObject with the gameObject having an associated enum to determine how it should be constrained.

Added EventConstraintMono and EventConstraintSO as wrappers for EventConstraints to allow use in the Inspector.

Added EventConstraintsEditor.

One of the options for the gameObject constraint is to determine if it is an instance of a given prefab. Added the following to support this:
  - PrefabReference component.
Gets automatically attatched to
all prefabs. It stores the prefab
reference and its lookup ID.
  - PrefabReferenceSO. Stores all prefab references and their associated lookup IDs. Returns the prefab reference when supplied an ID. Makes sure an instance of it exists at the given path. Automatically adds the PrefabReference component to all prefabs upon being created.
  - PrefabReferencesAssetPostprocessor. This notifies the PrefabReferenceSO when a prefab is created or deleted.

Changed the Event class to store its subscribers in a list rather than a UnityEvent. This is so it can be looped through and their constraints checked to determine if they should be called. This is apparently more efficient as well.

Added the GameState ScriptableObject which will be used in a state machine to determine the game state. These are used in EventConstraints.

This is all currently WIP and does not function correctly.