Update
Changed how information for the GameObjectConstraintType.InstanceOfPrefab constraint is stored. Now only stores a GUID string on the generated PrefabReference component.

Removed PrefabReferenceEditor as it is not longer needed.

Added BuildConsoleGUI to display console messages during play mode and builds.

Renamed PrefabReferences to PrefabReferencesManager to make it clearer what it does.

Changed the PrefabReferences component to only have a string field that contains a GUID string.

Removed PrefabReferenceSO since it is now unnecessary to store relations between prefabs, their given ID and their path.

Event now stores a List of actions with their constraints so they can be checked before being invoked.

Added a public method Invoke() in Event to be assigned in the Inspector if no EventInfo is required; passing an empty object would cause errors.