Update
Added tests for constraints.

Changed the constraints editor into a property drawer to display information in a nicer way when the constraints are a field (e.g. in constrains mono).

Renamed the EventConstraintsEditor class to EventConstraintsPropertyDrawer to match the changes to it.

Changed the gameobject constraints into a list in the Event class to allow for multiple constraints.

Commented out unimplemented GameObjectConstraintTypes.

Added a component menu attribute to the constraints mono class to make it clearer to designers that it is just a component for event constraints, they don't need to know there's multiple classes of them.

Renamed the PrefabReferences class to PrefabReferencesManager to make it clearer what it does.