Update
EventConstraints now works with GameState.

Added a GameStateManager component to set and revert the game state and attached it to an empty GameObject so it can be referenced from the scene.

Added a property drawer for EventConstraints to work with selecting desired GameStates from a multi choice menu to make it very easy to choose the desired constraints.

Added a utilities class for property drawers to make it much easier to lay them out. Limited functionality at the moment.

An AssetPostProcessor has been made to keep a constantly updated list of all GameState assets for the property drawer to use.

EventConstraints now correctly works with GameObjectConstraints and now takes a list of them.

Updated the scene MonoInstaller.

Added a references class that Zenject provides references to, so ScriptableObjects can access needed references.

The Event class now correctly resets it state on entering play mode so that actions aren't subscribed more times than they should.

Added more parameters for subscribing to an Event to allow for multiple subscriptions of the same action to be unique and possibly override existing ones.

Removed unneeded GameObjectConstraintTypes.

Initialised some empty Lists in various files to avoid NullReferenceExceptions.

Added an EventListener component to make it easy for GameObjects to respond to events.

Added unit tests. Also added Assembly definitions to make them work.

Installed the Guard API.