Add
Fixed NullReferenceException in EventListener when attempting to subscribe to Event with null constraints.

Added Description SO which is just a list of strings.

Added DescriptionReader component which takes a Description and a List<TextMeshPro>. It can display/conceal the description by looping through the list and setting the text value to the corresponding string in the description list, if it exists.

Added a panel to display information about a selected Building, Unit or Tile.

Added a series of panels to display player information: name and resources currently.

Installed TMPro.

Fixed a bug in the EventConstraintsGameObjectTests where a MonoBehaviour was being created with new().

Added GameObjectConstraintType.HasComponent.

Remove guard statement in Event that checks if constraint GameObject is null since not all constraints require a GameObject.

Added Unit class and imported previously made unit assets.

Changed TileType to Resource to make it clearer what it is.

Installed the new Input System.

Changed InteractableSelected Event to GameObjectSelected Event.

Added GameObjectDeselected Event.

Added InputManager. Gets the GameObject that was clicked by the mouse (clicked and released on) (doesn't go through UI) and invokes the GameObjectDeselected Event if the object clicked was not the currently selected GameObject, then invokes the GameObjectSelected Event.

Moved the GetTileClosestToMouse method from Town to InputManager.

Fixed IndexOutOfRange error in the TryGetTileAtPosition method of BoardManager.

Added OptionsMessage to display a message in the centre of the screen with an arbitrary number of options (buttons) below it.