Refactor TileBorders
Removed cubed loop.

Removed BorderInfo class as it was unnecessarily storing data; the data is only being used once and can be retrieved when it is to be used.

Changed the Side enum to Sides class which makes the code to assign and store which side to draw a border on shorter and easier to read.