Fix tile scale and border size
Fixed Tile scale and rotation in prefab.

Removed border scale adjustment since the tile scale is now correct.

Fixed border bottom left tile position. Was being clamped to a minimum of 0 so the border was being drawn too large if the tile should have been off the board.

Removed "Current commit changes.txt" as it was not meant to be committed.