Update
The border vfx is no longer rotated to match how the tile previously was.

The border vfx no longer resets on spawn.

TileBorder now only requires a local position and y rotation to spawn, making it better then having to calculate the exact position for both vertices each time.
This currently only works with square tiles but can be easily extended to work with other regular shapes; a fair amount of change will be required to make it work with irregular shapes.